# FleetManager/REST/models/command.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401
from typing import List, Dict, Any, Union  # noqa: F401  # CHANGED: Added Any and Union

from REST.models.base_model_ import Model
from REST.models.related_task import RelatedTask  # noqa: F401,E501
from REST import util

class Command(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, related_task: RelatedTask=None, id: int=None, command_type: str=None, command_status: str=None, start_time: int=None, params: List[Union[float, str, List[Union[str, float]]]]=None):  # noqa: E501  # CHANGED: params type
        """Command - a model defined in Swagger

        :param related_task: The related_task of this Command.  # noqa: E501
        :type related_task: RelatedTask
        :param id: The id of this Command.  # noqa: E501
        :type id: int
        :param command_type: The command_type of this Command.  # noqa: E501
        :type command_type: str
        :param command_status: The command_status of this Command.  # noqa: E501
        :type command_status: str
        :param start_time: The start_time of this Command.  # noqa: E501
        :type start_time: int
        :param params: The params of this Command.  # noqa: E501
        :type params: List[Union[float, str, List[Union[str, float]]]  # CHANGED
        """
        self.swagger_types = {
            'related_task': RelatedTask,
            'id': int,
            'command_type': str,
            'command_status': str,
            'start_time': int,
            'params': List[object]  # CHANGED: Use generic object type
        }

        self.attribute_map = {
            'related_task': 'relatedTask',
            'id': 'id',
            'command_type': 'commandType',
            'command_status': 'commandStatus',
            'start_time': 'startTime',
            'params': 'params'
        }
        self._related_task = related_task
        self._id = id
        self._command_type = command_type
        self._command_status = command_status
        self._start_time = start_time
        self._params = params

    @classmethod
    def from_dict(cls, dikt) -> 'Command':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Command of this Command.  # noqa: E501
        :rtype: Command
        """
        return util.deserialize_model(dikt, cls)

    @property
    def related_task(self) -> RelatedTask:
        """Gets the related_task of this Command.

        :return: The related_task of this Command.
        :rtype: RelatedTask
        """
        return self._related_task

    @related_task.setter
    def related_task(self, related_task: RelatedTask):
        """Sets the related_task of this Command.

        :param related_task: The related_task of this Command.
        :type related_task: RelatedTask
        """
        self._related_task = related_task

    @property
    def id(self) -> int:
        """Gets the id of this Command.

        :return: The id of this Command.
        :rtype: int
        """
        return self._id

    @id.setter
    def id(self, id: int):
        """Sets the id of this Command.

        :param id: The id of this Command.
        :type id: int
        """
        self._id = id

    @property
    def command_type(self) -> str:
        """Gets the command_type of this Command.

        :return: The command_type of this Command.
        :rtype: str
        """
        return self._command_type

    @command_type.setter
    def command_type(self, command_type: str):
        """Sets the command_type of this Command.

        :param command_type: The command_type of this Command.
        :type command_type: str
        """
        allowed_values = ["NAV_TAKEOFF", "NAV_WAYPOINT", "CAMERA_IMAGE", 
                          "ANALYZE_IMAGE", "NAV_HOME", "NAV_LAND", "SPRAY"]
        if command_type not in allowed_values:
            raise ValueError(
                f"Invalid command_type: '{command_type}'. Must be one of {allowed_values}"
            )
        self._command_type = command_type

    @property
    def command_status(self) -> str:
        """Gets the command_status of this Command.

        :return: The command_status of this Command.
        :rtype: str
        """
        return self._command_status

    @command_status.setter
    def command_status(self, command_status: str):
        """Sets the command_status of this Command.

        :param command_status: The command_status of this Command.
        :type command_status: str
        """
        # CHANGED: Expanded enum values
        allowed_values = ["NotAssigned", "Assigned", "Executing", "Completed", "Failed"]
        if command_status not in allowed_values:
            raise ValueError(
                f"Invalid command_status: '{command_status}'. Must be one of {allowed_values}"
            )
        self._command_status = command_status

    @property
    def start_time(self) -> int:
        """Gets the start_time of this Command.

        :return: The start_time of this Command.
        :rtype: int
        """
        return self._start_time

    @start_time.setter
    def start_time(self, start_time: int):
        """Sets the start_time of this Command.

        :param start_time: The start_time of this Command.
        :type start_time: int
        """
        self._start_time = start_time

    @property
    def params(self) -> List[Union[float, str, List[Union[str, float]]]]:
        """Gets the params of this Command.

        :return: The params of this Command.
        :rtype: List[Union[float, str, List[Union[str, float]]]]
        """
        return self._params

    @params.setter
    def params(self, params: List[Union[float, str, List[Union[str, float]]]]):
        """Sets the params of this Command.

        :param params: The params of this Command.
        :type params: List[Union[float, str, List[Union[str, float]]]]
        """
        # CHANGED: No type restriction beyond basic validation
        self._params = params