# FleetManager/REST/models/related_task.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from REST.models.base_model_ import Model
from REST import util


class RelatedTask(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, id: int=None, mission_id: int=None):  # noqa: E501
        """RelatedTask - a model defined in Swagger

        :param id: The id of this RelatedTask.  # noqa: E501
        :type id: int
        :param mission_id: The mission_id of this RelatedTask.  # noqa: E501
        :type mission_id: int
        """
        self.swagger_types = {
            'id': int,
            'mission_id': int
        }

        self.attribute_map = {
            'id': 'id',
            'mission_id': 'missionId'
        }
        self._id = id
        self._mission_id = mission_id

    @classmethod
    def from_dict(cls, dikt) -> 'RelatedTask':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The RelatedTask of this RelatedTask.  # noqa: E501
        :rtype: RelatedTask
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> int:
        """Gets the id of this RelatedTask.


        :return: The id of this RelatedTask.
        :rtype: int
        """
        return self._id

    @id.setter
    def id(self, id: int):
        """Sets the id of this RelatedTask.


        :param id: The id of this RelatedTask.
        :type id: int
        """

        self._id = id

    @property
    def mission_id(self) -> int:
        """Gets the mission_id of this RelatedTask.


        :return: The mission_id of this RelatedTask.
        :rtype: int
        """
        return self._mission_id

    @mission_id.setter
    def mission_id(self, mission_id: int):
        """Sets the mission_id of this RelatedTask.


        :param mission_id: The mission_id of this RelatedTask.
        :type mission_id: int
        """

        self._mission_id = mission_id