# FleetManager/REST/models/task_dependency.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from REST.models.base_model_ import Model
from REST import util


class TaskDependency(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, dependency_status: str=None, dependency_task_id: int=None, dependency_task_status: str=None, dependency_result_type: List[int]=None):  # noqa: E501
        """TaskDependency - a model defined in Swagger

        :param dependency_status: The dependency_status of this TaskDependency.  # noqa: E501
        :type dependency_status: str
        :param dependency_task_id: The dependency_task_id of this TaskDependency.  # noqa: E501
        :type dependency_task_id: int
        :param dependency_task_status: The dependency_task_status of this TaskDependency.  # noqa: E501
        :type dependency_task_status: str
        :param dependency_result_type: The dependency_result_type of this TaskDependency.  # noqa: E501
        :type dependency_result_type: List[int]
        """
        self.swagger_types = {
            'dependency_status': str,
            'dependency_task_id': int,
            'dependency_task_status': str,
            'dependency_result_type': List[int]
        }

        self.attribute_map = {
            'dependency_status': 'dependencyStatus',
            'dependency_task_id': 'dependencyTaskId',
            'dependency_task_status': 'dependencyTaskStatus',
            'dependency_result_type': 'dependencyResultType'
        }
        self._dependency_status = dependency_status
        self._dependency_task_id = dependency_task_id
        self._dependency_task_status = dependency_task_status
        self._dependency_result_type = dependency_result_type

    @classmethod
    def from_dict(cls, dikt) -> 'TaskDependency':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The TaskDependency of this TaskDependency.  # noqa: E501
        :rtype: TaskDependency
        """
        return util.deserialize_model(dikt, cls)

    @property
    def dependency_status(self) -> str:
        """Gets the dependency_status of this TaskDependency.


        :return: The dependency_status of this TaskDependency.
        :rtype: str
        """
        return self._dependency_status

    @dependency_status.setter
    def dependency_status(self, dependency_status: str):
        """Sets the dependency_status of this TaskDependency.


        :param dependency_status: The dependency_status of this TaskDependency.
        :type dependency_status: str
        """
        allowed_values = ["NotModified"]  # noqa: E501
        if dependency_status not in allowed_values:
            raise ValueError(
                "Invalid value for `dependency_status` ({0}), must be one of {1}"
                .format(dependency_status, allowed_values)
            )

        self._dependency_status = dependency_status

    @property
    def dependency_task_id(self) -> int:
        """Gets the dependency_task_id of this TaskDependency.


        :return: The dependency_task_id of this TaskDependency.
        :rtype: int
        """
        return self._dependency_task_id

    @dependency_task_id.setter
    def dependency_task_id(self, dependency_task_id: int):
        """Sets the dependency_task_id of this TaskDependency.


        :param dependency_task_id: The dependency_task_id of this TaskDependency.
        :type dependency_task_id: int
        """
        if dependency_task_id is None:
            raise ValueError("Invalid value for `dependency_task_id`, must not be `None`")  # noqa: E501

        self._dependency_task_id = dependency_task_id

    @property
    def dependency_task_status(self) -> str:
        """Gets the dependency_task_status of this TaskDependency.


        :return: The dependency_task_status of this TaskDependency.
        :rtype: str
        """
        return self._dependency_task_status

    @dependency_task_status.setter
    def dependency_task_status(self, dependency_task_status: str):
        """Sets the dependency_task_status of this TaskDependency.


        :param dependency_task_status: The dependency_task_status of this TaskDependency.
        :type dependency_task_status: str
        """
        allowed_values = ["Finished"]  # noqa: E501
        if dependency_task_status not in allowed_values:
            raise ValueError(
                "Invalid value for `dependency_task_status` ({0}), must be one of {1}"
                .format(dependency_task_status, allowed_values)
            )

        self._dependency_task_status = dependency_task_status

    @property
    def dependency_result_type(self) -> List[int]:
        """Gets the dependency_result_type of this TaskDependency.


        :return: The dependency_result_type of this TaskDependency.
        :rtype: List[int]
        """
        return self._dependency_result_type

    @dependency_result_type.setter
    def dependency_result_type(self, dependency_result_type: List[int]):
        """Sets the dependency_result_type of this TaskDependency.


        :param dependency_result_type: The dependency_result_type of this TaskDependency.
        :type dependency_result_type: List[int]
        """
        if dependency_result_type is None:
            raise ValueError("Invalid value for `dependency_result_type`, must not be `None`")  # noqa: E501

        self._dependency_result_type = dependency_result_type