# FleetManager/REST/models/vehicle.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from REST.models.base_model_ import Model
from REST import util


class Vehicle(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, id: int=None, type: str=None):  # noqa: E501
        """Vehicle - a model defined in Swagger

        :param id: The id of this Vehicle.  # noqa: E501
        :type id: int
        :param type: The type of this Vehicle.  # noqa: E501
        :type type: str
        """
        self.swagger_types = {
            'id': int,
            'type': str
        }

        self.attribute_map = {
            'id': 'id',
            'type': 'type'
        }
        self._id = id
        self._type = type

    @classmethod
    def from_dict(cls, dikt) -> 'Vehicle':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Vehicle of this Vehicle.  # noqa: E501
        :rtype: Vehicle
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> int:
        """Gets the id of this Vehicle.


        :return: The id of this Vehicle.
        :rtype: int
        """
        return self._id

    @id.setter
    def id(self, id: int):
        """Sets the id of this Vehicle.


        :param id: The id of this Vehicle.
        :type id: int
        """

        self._id = id

    @property
    def type(self) -> str:
        """Gets the type of this Vehicle.


        :return: The type of this Vehicle.
        :rtype: str
        """
        return self._type

    @type.setter
    def type(self, type: str):
        """Sets the type of this Vehicle.


        :param type: The type of this Vehicle.
        :type type: str
        """
        allowed_values = ["AUAV"]  # noqa: E501
        if type not in allowed_values:
            raise ValueError(
                "Invalid value for `type` ({0}), must be one of {1}"
                .format(type, allowed_values)
            )

        self._type = type